Bugfix: mass actions required when we have more than one element
git-svn-id: file:///home/code/svn/ovd/trunk@2214 2b2718be-6ee5-4ec5-ad1a-3fa14f9c7ca3

+++ SessionManager/web/admin/sessions.php
@@ -192,7 +192,8 @@ else {
 	if (is_array($sessions) && count($sessions) > 0) {
 		echo '<table class="main_sub sortable" id="sessions_list_table" border="0" cellspacing="1" cellpadding="3">';
 		echo '	<tr class="title">';
-		echo '		<th class="unsortable"></th>';
+		if (count($sessions) > 1)
+			echo '		<th class="unsortable"></th>';
 		echo '		<th>'._('Session').'</th>';
 		echo '		<th>'._('Server').'</th>';
 		echo '		<th>'._('User').'</th>';
@@ -204,7 +205,8 @@ else {
 			$css_class = 'content'.(($i++%2==0)?1:2);
 
 			echo '	<tr class="'.$css_class.'">';
-			echo '		<td><input class="input_checkbox" type="checkbox" name="kill_sessions[]" value="'.$session->id.'" /></td>';
+			if (count($sessions) > 1)
+				echo '		<td><input class="input_checkbox" type="checkbox" name="kill_sessions[]" value="'.$session->id.'" /></td>';
 			echo '		<td><a href="sessions.php?info='.$session->id.'">'.$session->id.'</td>';
 			echo '		<td><a href="servers.php?action=manage&fqdn='.$session->server.'">'.$session->server.'</td>';
 			echo '		<td><a href="users.php?action=manage&id='.$session->getAttribute('user_login').'">'.$session->getAttribute('user_displayname').'</td>';
@@ -219,16 +221,18 @@ else {
 			echo '	</tr>';
 		}
 		$css_class = 'content'.(($i++%2==0)?1:2);
-		echo '<tfoot>';
-		echo '	<tr class="'.$css_class.'">';
-		echo '		<td colspan="5"><a href="javascript:;" onclick="markAllRows(\'sessions_list_table\'); return false">'._('Mark all').'</a> / <a href="javascript:;" onclick="unMarkAllRows(\'sessions_list_table\'); return false">'._('Unmark all').'</a></td>';
-		echo '<td>';
-		echo '<form action="sessions.php" method="post" onsubmit="return confirm(\''._('Are you sure you want to kill selected sessions?').'\') && updateMassActionsForm(this, \'sessions_list_table\');">';
-		echo '	<input type="hidden" name="mass_action" value="kill" />';
-		echo '<input type="submit" name="kill" value="'._('Kill').'" />';
-		echo '</td>';
-		echo '	</tr>';
-		echo '</tfoot>';
+		if (count($sessions) > 1) {
+			echo '<tfoot>';
+			echo '	<tr class="'.$css_class.'">';
+			echo '		<td colspan="5"><a href="javascript:;" onclick="markAllRows(\'sessions_list_table\'); return false">'._('Mark all').'</a> / <a href="javascript:;" onclick="unMarkAllRows(\'sessions_list_table\'); return false">'._('Unmark all').'</a></td>';
+			echo '<td>';
+			echo '<form action="sessions.php" method="post" onsubmit="return confirm(\''._('Are you sure you want to kill selected sessions?').'\') && updateMassActionsForm(this, \'sessions_list_table\');">';
+			echo '	<input type="hidden" name="mass_action" value="kill" />';
+			echo '<input type="submit" name="kill" value="'._('Kill').'" />';
+			echo '</td>';
+			echo '	</tr>';
+			echo '</tfoot>';
+		}
 		echo '</table>';
 	} else {
 		echo _('No active session');
